requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==14.2.0
django-heroku==0.3.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
Pillow==8.1.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-decouple==3.4
pytz==2020.5
requests==2.25.1
rsa==4.7
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
